requirements.txt
#amqp==2.1.4
#backports.zoneinfo==0.2.1
#billiard==3.5.0.5
#celery==4.1.1
Django==1.10.2
django-braces==1.14.0
django-celery-beat==1.1.1
django-ckeditor==6.1.0
django-cors-headers==1.2.1
django-datatables-view==1.14.0
django-excel==0.0.10
django-js-asset==1.2.2
django-oauth-toolkit==0.12.0
djangorestframework==3.4.7
#et-xmlfile==1.1.0
#importlib-metadata==4.0.1
#importlib-resources==5.4.0
#kombu==4.6.11
#lml==0.1.0
#numpy==1.19.5
oauthlib==2.0.1
#openpyxl==3.0.9
#pandas==1.1.5
#pdfkit==0.5.0
pkg-resources==0.0.0
psycopg==3.0.10
psycopg2==2.9.3
psycopg2-binary==2.9.3
pycrypto==2.6.1
#pyexcel==0.6.6
#pyexcel-io==0.6.4
#pyexcel-webio==0.1.4
#python-dateutil==2.7.3
#pytz==2021.1
#six==1.16.0
#texttable==1.6.3
#typing-extensions==4.1.1
#vine==1.3.0
#zipp==3.4.1
